+++ src/main/java/io/kyberorg/yalsee/ui/MainView.java
@@ -5,6 +5,7 @@ import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.applayout.AppLayout;
 import com.vaadin.flow.component.applayout.DrawerToggle;
 import com.vaadin.flow.component.dependency.CssImport;
+import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.component.html.Anchor;
 import com.vaadin.flow.component.html.Image;
 import com.vaadin.flow.component.html.Span;
@@ -56,6 +57,7 @@ import static io.kyberorg.yalsee.ui.MainView.IDs.APP_LOGO;
 @SpringComponent
 @UIScope
 @CssImport("./css/main_view.css")
+@JsModule("./js/open-share-menu.js")
 public class MainView extends AppLayout implements BeforeEnterObserver {
     private static final String TAG = "[" + MainView.class.getSimpleName() + "]";