+++ src/main/java/io/kyberorg/yalsee/ui/pages/home/HomePage.java
@@ -6,6 +6,7 @@ import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.button.ButtonVariant;
 import com.vaadin.flow.component.dependency.CssImport;
 import com.vaadin.flow.component.html.*;
+import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.component.notification.Notification;
 import com.vaadin.flow.component.orderedlayout.FlexComponent;
@@ -34,10 +35,9 @@ import io.kyberorg.yalsee.services.QRCodeService;
 import io.kyberorg.yalsee.services.overall.OverallService;
 import io.kyberorg.yalsee.session.YalseeSession;
 import io.kyberorg.yalsee.ui.MainView;
-import io.kyberorg.yalsee.utils.AppUtils;
-import io.kyberorg.yalsee.utils.ClipboardUtils;
-import io.kyberorg.yalsee.utils.ErrorUtils;
-import io.kyberorg.yalsee.utils.UrlUtils;
+import io.kyberorg.yalsee.ui.elements.MobileShareMenu;
+import io.kyberorg.yalsee.ui.elements.ShareMenu;
+import io.kyberorg.yalsee.utils.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
@@ -50,6 +50,7 @@ import org.greenrobot.eventbus.Subscribe;
 @UIScope
 @CssImport("./css/common_styles.css")
 @CssImport("./css/home_view.css")
+
 @Route(value = Endpoint.UI.HOME_PAGE, layout = MainView.class)
 @PageTitle("Yalsee - the link shortener")
 public class HomePage extends HorizontalLayout implements BeforeEnterObserver {
@@ -84,6 +85,9 @@ public class HomePage extends HorizontalLayout implements BeforeEnterObserver {
     private Span linkCounter;
 
     private Notification errorNotification;
+    private ShareMenu shareMenu;
+
+    private String descriptionInputHolder;
 
     @Override
     public void beforeEnter(final BeforeEnterEvent beforeEnterEvent) {
@@ -100,6 +104,8 @@ public class HomePage extends HorizontalLayout implements BeforeEnterObserver {
 
         removeAll();
         add(leftDiv, centralLayout, rightDiv);
+
+        shareMenu = ShareMenu.create();
     }
 
     private void applyStyle() {
@@ -204,15 +210,19 @@ public class HomePage extends HorizontalLayout implements BeforeEnterObserver {
         shortLink.setId(IDs.SHORT_LINK);
         shortLink.addClassName("strong-link");
 
-        com.vaadin.flow.component.icon.Icon copyLinkImage;
-        copyLinkImage = new com.vaadin.flow.component.icon.Icon(VaadinIcon.COPY);
+        Icon shareIcon = new Icon(VaadinIcon.SHARE);
+        shareIcon.setId(IDs.SHARE_ICON);
+        shareIcon.addClickListener(this::openShareMenu);
+
+        Icon copyLinkImage;
+        copyLinkImage = new Icon(VaadinIcon.COPY);
         copyLinkImage.setId(IDs.COPY_LINK_BUTTON);
         copyLinkImage.addClickListener(this::copyLinkToClipboard);
 
         resultArea.setJustifyContentMode(FlexComponent.JustifyContentMode.CENTER);
         resultArea.setDefaultVerticalComponentAlignment(Alignment.CENTER);
 
-        resultArea.add(emptySpan, shortLink, copyLinkImage);
+        resultArea.add(emptySpan, shortLink, shareIcon, copyLinkImage);
         resultArea.addClassNames("result-area", "border");
         resultArea.setWidthFull();
         return resultArea;
@@ -340,6 +350,7 @@ public class HomePage extends HorizontalLayout implements BeforeEnterObserver {
         }
 
         if (isFormValid) {
+            saveDescription();
             cleanForm();
             cleanResults();
             saveLink(longUrl, linkDescription);
@@ -463,6 +474,10 @@ public class HomePage extends HorizontalLayout implements BeforeEnterObserver {
         errorNotification.open();
     }
 
+    private void saveDescription() {
+        descriptionInputHolder = descriptionInput.getValue();
+    }
+
     private void cleanForm() {
         input.setValue("");
         protocolSelector.setVisible(false);
@@ -512,6 +527,19 @@ public class HomePage extends HorizontalLayout implements BeforeEnterObserver {
         }
     }
 
+    private void openShareMenu(final ClickEvent<Icon> iconClickEvent) {
+        log.trace("{} Share menu clicked. From client? {}", TAG, iconClickEvent.isFromClient());
+        if (ui != null && ui.getPage() != null && DeviceUtils.isMobileDevice()) {
+            MobileShareMenu.createForPage(ui.getPage()).setLink(shortLink.getText()).show();
+        } else {
+            shareMenu.setShortLink(shortLink.getText());
+            if (StringUtils.isNotBlank(descriptionInputHolder)) {
+                shareMenu.setDescription(descriptionInputHolder);
+            }
+            shareMenu.show();
+        }
+    }
+
     public static class IDs {
         public static final String VIEW_ID = "homeView";
         public static final String MAIN_AREA = "mainArea";
@@ -540,5 +568,6 @@ public class HomePage extends HorizontalLayout implements BeforeEnterObserver {
         public static final String MY_LINKS_NOTE_START = "myLinksNoteStart";
         public static final String MY_LINKS_NOTE_LINK = "myLinksNoteLink";
         public static final String MY_LINKS_NOTE_END = "myLinksNoteEnd";
+        public static final String SHARE_ICON = "shareMenu";
     }
 }